split files and cleaned trunc
